configure mainnet using build tags
